マーケティングレイアウトとページの追加、不要なページの削除
