+++ src/app/(marketing)/pricing/page.tsx
